+++ src/main/java/Table/Entry.java
@@ -1,23 +1,23 @@
-package Table;
-
-public class Entry<K, V> {
-    private K key;
-    private V value;
-
-    public Entry(K key, V value) {
-        this.key = key;
-        this.value = value;
-    }
-
-    public K getKey() {
-        return key;
-    }
-
-    public V getValue() {
-        return value;
-    }
-
-    public void setValue(V value) {
-        this.value = value;
-    }
-}
+//package Table;
+//
+//public class Entry<K, V> {
+//    private K key;
+//    private V value;
+//
+//    public Entry(K key, V value) {
+//        this.key = key;
+//        this.value = value;
+//    }
+//
+//    public K getKey() {
+//        return key;
+//    }
+//
+//    public V getValue() {
+//        return value;
+//    }
+//
+//    public void setValue(V value) {
+//        this.value = value;
+//    }
+//}
