+++ src/main/java/Table/Table.java
@@ -10,6 +10,27 @@ import java.util.ArrayList;
  * Void return on `remove`.
  */
 public class Table<K, V> {
+    public class Entry {
+        private K key;
+        private V value;
+
+        public Entry(K key, V value) {
+            this.key = key;
+            this.value = value;
+        }
+
+        public K getKey() {
+            return key;
+        }
+
+        public V getValue() {
+            return value;
+        }
+
+        public void setValue(V value) {
+            this.value = value;
+        }
+    }
     private ArrayList<Entry> entries;
 
     public Table() {
